Update patch to include a bug fix pointed out by Daniel Hartmeier This looks like a bug in pf_ioctl.c pfioctl() DIOCCHANGERULE
                       if (((((newrule->action == PF_NAT) ||
                           (newrule->action == PF_RDR) ||
                           (newrule->action == PF_BINAT) ||
                           (newrule->rt > PF_FASTROUTE)) &&
-                           !pcr->anchor[0])) &&
+                           !newrule->anchor)) &&
                           (TAILQ_FIRST(&newrule->rpool.list) == NULL))
                               error = EINVAL;

i.e. the pool must not be empty for routing and translation rules,
except for translation rules that are actually anchor _calls_.

The confusion is between translation rules within anchors
(pcr->anchor[0] != '\0') and calls to anchors' translation rules
(rule->anchor != NULL).

If the proxy is using DIOCCHANGERULE (it must be the proxy, pfctl isn't
using it at all), AND is trying to add/update a rule that requires at
least one replacement address but contains an empty list, then this
would cause the panic seen when that rule later matches a packet.

This needs fixing in OpenBSD as well.