Fix issues with packet recursing in the stack when dummynet(4), ipfw(4) and pf(4) are loaded.
This allows us to reenable the shaping code for the captive portal to be feature par
with m0n0 at the captive portal side.

Beaware that dummynet(4) and pfSense multi-wan are not compatible.
Ipfw(4) needs to be enabled before pf(4) module to allow that but as things work
for now there is no way to achive this. Maybe, find some time to allow pfil(9) hooks
to be reordered so shaping with dummynet(4) and multi-wan are compatible.

TODO: Import dummynet(4) patches for pf(4) and export the option to the GUI.

I think this commit alone would make a lot of Captive Portal consumers happy!