If /usrport is found in the pfPort directory then do not remove the target port directory.
This is handy if we want to track a port but do not care about it changing over time and
will allow us to need to keep the port up to date in pfPorts.  It will simply build the
found port from /usr/ports/* and use it without modifying it.