Add install MBR option.  From David Samms:
When installing the "embedded kernel" on to a hard disk drive the correct mbr is written to /boot/mbr, but it is NOT installed with fdisk.  To recover, I reboot of the CD a second time, copy /boot/mbr off the hard disk, and then run fdisk -B -b mbr ad0.  If this is not done, the ALIX won't boot off of the hard disk, instead, you get a prompt asking for "Default:   F1".