Patch from Raj:
1. Per Build config file: This patch allows for a per build config
file. Per build variables are specified in a file and the file is
passed as an argument to build command. for example

./build_nano.sh custom_cfg/kvm.conf

Where custom_cfg/kvm.conf consists of

export USE_CONFIG_XML=${USE_CONFIG_XML:-"custom_cfg/kvm.xml"}

Useful to specify a separate conf file for kvm build. Any build
variable can be specified using this method.

Currently patch is for build_nano.sh, but this is useful for all builds.