Adding patch to correct:
Import change from KAME, rev. 1.362 kame/kame/sys/netkey/key.c

In case of "new SA", we must check the hard lifetime of the old SA
to find out if it is not permanent and we can delete it.

Submitted by:	sakane via gnn
MFC after:	3 days

http://svn.freebsd.org/viewvc/base?view=revision&revision=177553

Thanks to Bjoern A. Zeeb for pointing it out