Adding fixes from Luigi Rizzo:
+ make 'noupdate' the default mode of operation (this is
 forced in the Makefile). It can be reverted back to 'update'
 using boot0cfg.

+ never overwrite the boot sector if the BIOS-supplied drive
 number is overridden by the 'setdrv' option. This was a
 potential source of trouble because we might write (and trash
 the MBR) on a different driver.

+ do not check for a valid drive number, allowing boot0 to be used
 even when the BIOS does 'floppy emulation' on a flash drive.
 Adding the check back requires 4 bytes.

+ force CHS mode ('nopacket') when loaded from a floppy unit.
 Removing the change saves 4 bytes.

Because of some minor code rearrangements, I also managed to make
room for a 'WIN' string (3 extra bytes) to be printed for certain
FAT32 partitions, and for recognising an additional FAT16 partition
as DOS (2 extra bytes).

Right now the code uses the full 512 bytes (the SIO version
is slightly shorter) so the changes need to fit in the memory
budget -- this is why i mention the cost of each of the options above.