Fix CP disapearing packets problem introduced with the fixes to allow CP to work on multi-wan.
The problem is that the check for ipfw forwarded packets is done to early and since pf is a stateful firewall it needs to update it state data for each packet that traverses the firewall.
The fix consists on moving the check after pf has done its data collection and forbid it from doing any route-to/reply-to magic so CP and multi-wan can work.