Adding patch to resolve tftp-proxy issues:
"Today there was a need to have tftp-proxy, but I found that it does not work.

It was this way:
Jul 6 10:53:27 vakhtyor tftp-proxy [47783]: pf connection lookup failed (no rdr?)
Jul 6 10:53:27 vakhtyor inetd [47778]: / usr / libexec / tftp-proxy [47783]: exited, status 1

After that I looked the source code, and found the error:
tftp-proxy trying to find is state connection to him and did not find it, because not correctly sets the protocol:

server_lookup4 (struct sockaddr_in * client, struct sockaddr_in * proxy,
   struct sockaddr_in * server)
(
.....
-->         pnl.proto = IPPROTO_TCP;
.....

After that I started looking for where the legs grow and discovered that it corrected openbsd (added argument function proto).
It may be to synchronize the source code?"