patches/RELENG_8_3: ng_tcpmss.diff removal
This patch archive wasn't enabled for builds anyway. Checking history
shows this patch was introduced into builds based on FreeBSD 8.1, but
within then later been disabled because of regressions caused by it.

pfSense issue #1336, comment 14:

"[...] the patch had to be backed out for now - it appears to have been
negatively impacting pppoe on amd64.

This