patches/RELENG_8_3: esx-carp.diff removal
This patch hasn't been enabled in our builds for long
time (if ever according to logs).

Digging archives reveals that this patch was built by shrew.net for a
bug with CARP when running FreeBSD with CARP on VMware ESX(i) 3.5 and
4.0. VMware fixed this issue on their side with ESX(i) 4.0 Update 2.

Thus I see no need to carry that patch with us any longer.

For references to VMware bug fix:
http://www.vmware.com/support/vsphere4/doc/vsp_esxi40_u2_rel_notes.html