patches/RELENG_8_3: SR9600.diff removal
This patch has been disabled for 8.3 based builds. Looking further
revealed that Sunrising SR9600 has the exact same vendor and hardware
ID of Kontron AG DM9601 which is already present in 8.3.

I guess it's just a matter of a differentely branded boxing of the
physical hardware but the chips must be the same.

Additionally to the mess Linux, like FreeBSD says the vendor ID is
Kontron while NetBSD says its Sunrising...