patches/RELENG_8_3: Import bge and bce backport from 8-STABLE
This is a huge commit, I admit, but it adds quite some bugfixes
taken from 8-STABLE related to the widely used NetExtreme and
NetExtreme II server NICs that might be handy.

In the hope that it helps to keep pfSense 2.1 up-to-date via 8-STABLE
for more recent hardware since 8.3's release.

Agreed: I import quite a number of patches after wiping others them...