pfPorts: expiretable, replace / bump w. FreeBSD ports r289156
This was a local port by pfSense, not updated since 2005 and had
a couple of typo and other style errors. Additionally it contained
the whole source of expiretable in our codebase.

FreeBSD has an up-to-date and maintained expiretables, use theirs from now on.

Checking in-tree source showed that we had an unmodified expiretable 0.4 in use.
New port is now based on this upstream commit:

Author: dougb <dougb@FreeBSD.org>
Date:   Sat Jan 14 08:57:23 2012 +0000

    In the rc.d scripts, change assignments to rcvar to use the
    literal name_enable wherever possible, and ${name}_enable
    when it's not, to prepare for the demise of set_rcvar().

    In cases where I had to hand-edit unusual instances also
    modify formatting slightly to be more uniform (and in
    some cases, correct). This includes adding some $FreeBSD$
    tags, and most importantly moving rcvar= to right after
    name= so it's clear that one is derived from the other.