pfPorts: Remove libmrss 0.13
This port has not been update since 2006, it was is found in pkg_config.xml
in use by phpmrss package, but nowhere else. Not touch within the last 6 years.

The libmrss port is still available in FreeBSD ports if we need to resurrect it.
Additionally to note, libmrss code not seen any changes since 2007.

Actually some of our binaries on files.pfsense.org contain have 19.1, not
this deleted hereby, this is a bit confusing.