pfPorts: Sync unbound 1.4.19 (FreeBSD r308903), disable ECDSA
Sync with FreeBSD ports for a regular unbound release, try to stay
as close as possible.

We have to disable ECDSA support until we ship OpenSSL 1.0 as ECDSA
requires this newer OpenSSL version not present in base FreeBSd 8.3.
The upstream project denotes it 0.9.8 would work, but has known issues.
(Reference: unbound 1.4.17 release notes)