pfPorts/php53 sync [3/3]: Rely on bsdtar format auto-determination (FreeBSD Ports r316786)
Upstream message:
Rely on bsdtar to autodetermine the format of the distfiles when possible

For a while now bsdtar is able to autotermine compression and archive format.
Let's then use tar directly instead of piping to tar.

Now USE_BZIP2 and USE_XZ only set EXTRACT_SUFX to the right format

pfSense notes:
- Our patches still seem to apply correctly.