Add filterlog port that will help to make filterlogs parsing and display easily.
Most of the parsing code is adapted from tcpdump since the pcap part is standard allover.
Tweaks have been done to the tcpdump to display the info in a parsable way and some details not needed.

For now TCP/UDP/CARP and ipv4/ipv6 parsing is done, for other protos its just a question of putting the info needed.